Fix license key in Cargo.toml
The Cargo manifest format requires a specific syntax for specifying the license key. `MIT/Apache-2.0` is invalid in that syntax, so some tools might not recognize the license. You can read more about the syntax in [the Cargo reference format docs](https://doc.rust-lang.org/cargo/reference/manifest.html#the-license-and-license-file-fields).

Instead, we now use `MIT OR Apache-2.0`, which is the intended way to write this license.